close #2

+++ src/main/java/it/unipd/tos/Model/MenuItem.java
@@ -8,7 +8,7 @@ public class MenuItem {
     private ItemType itemType;
     private String name;
     private double price;
-    public MenuItem(ItemType itemType, String name, int price) {
+    public MenuItem(ItemType itemType, String name, double price) {
         this.itemType = itemType;
         this.name = name;
         this.price = price;
